+++ .gitignore
@@ -195,3 +195,7 @@ modules.order
 Module.symvers
 Mkfile.old
 dkms.conf
+
+# Custom
+.vscode/*
+issues
