+++ include/afc_helpers.h
@@ -1,9 +1,11 @@
-#pragma once
+#ifndef AFC_HELPERS_H
+#define AFC_HELPERS_H
+#include <string>
+
 #include <libimobiledevice/afc.h>
 #include <libimobiledevice/libimobiledevice.h>
 #include <libimobiledevice/lockdown.h>
 
-#include <string>
 /**
  * Build the full path for a directory entry.
  */
@@ -18,3 +20,5 @@ bool is_directory(afc_client_t afc, const char* fullPath);
  * Check if the file extension indicates an image file.
  */
 bool is_image_file(const std::string& filePath);
+
+#endif
