+++ src/afc_helpers.cpp
@@ -1,10 +1,9 @@
-// afc_helpers.cpp
-#include "afc_helpers.h"
-
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
 #include <string>
+
+#include "afc_helpers.h"
 
 char* build_full_path(const char* directory, const char* entry) {
     char* fullPath = nullptr;
